weight-decay needs to be float, not int

+++ galaxy_merge_edits/no_domain_adaptation.py
@@ -290,7 +290,7 @@ if __name__ == "__main__":
     parser.add_argument('--lr_scan', type=str, default = 'no', help='Set to yes for learning rate scan')
     parser.add_argument('--cycle_length', type=int, default = 2, help = 'If using one-cycle learning, how many epochs should one learning rate cycle be?')
     parser.add_argument('--early_stop_patience', type=int, default = 10, help = 'Number of epochs for early stopping.')
-    parser.add_argument('--weight_decay', type=int, default = 5e-4, help= 'How much do you want to penalize large weights?')
+    parser.add_argument('--weight_decay', type=float, default = 5e-4, help= 'How much do you want to penalize large weights?')
 
     args = parser.parse_args()
     os.environ["CUDA_VISIBLE_DEVICES"] = args.gpu_id
